requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
scipy>=1.11.0
statsmodels>=0.14.0
numpy>=1.24.0
